Arranger hack
Difficult to explain at the moment - when printing stuff that would
overflow to a new page if the first line had a soft-hyphen at the end
the part before that soft-hyphen would be unnecessarily moved to the
second line before when repacking unretrieved fragments we would return
text *without* hyphens and previous_fragment_ended_with_breakable? would
not match.